backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
transformers==4.35.2
torch==2.1.1
pinecone-client==2.2.4
pytesseract==0.3.10
Pillow==10.1.0
langdetect==1.0.9
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
googletrans==4.0.0
pydantic-settings==2.0.3
alembic==1.12.1
httpx==0.25.2